intに変えられたぜ！

+++ chapter02/sumMatrixOnGPU-2D-grid-2D-block.cu
@@ -11,7 +11,7 @@
  * iterates through vector elements on the host.
  */
 
-using T=float;
+using T=int;
 
 double cpuSecond(){
     struct timeval tp;
@@ -65,7 +65,7 @@ void checkResult(T *hostRef, T *gpuRef, const int N)
         if (abs(hostRef[i] - gpuRef[i]) > epsilon)
         {
             match = 0;
-            printf("host %f gpu %f\n", hostRef[i], gpuRef[i]);
+            printf("host %d gpu %d\n", hostRef[i], gpuRef[i]);
             break;
         }
     }
